Changed all refernences to new files.
