+++ src/features/vote/helpers/formatVoteDataToCommit.ts
@@ -1,6 +1,6 @@
 import web3 from "web3";
-import { PostCommitVote } from "web3/postVotingContractMethods";
-import { PendingRequest } from "web3/queryVotingContractMethods";
+import { PostCommitVote } from "web3/post/commitVotes";
+import { PendingRequest } from "web3/get/queryGetPendingRequests";
 import {
   computeVoteHashAncillary,
   getRandomSignedInt,
